wsi_service_base_plugins/pil/pyproject.toml
[tool.poetry]
name = "wsi-service-plugin-pil"
version = "0.3.0"
description = ""
authors = ["EMPAIA <dev-support@empaia.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
